Refactor game memory library into its own target
