+++ cmake/generate_sources.cmake
@@ -1,25 +1,33 @@
-# generate_sources.cmake — Run with: cmake -P generate_sources.cmake -DSOURCE_DIR=...
-# This script generates src/client/sources.cmake *only if* the contents have changed.
+# generate_sources.cmake — Run with: cmake -P generate_sources.cmake -DSOURCE_DIR=... -DTARGET_NAME=... -DSEARCH_DIR=...
+# This script generates src/{target}/sources.cmake *only if* the contents have changed.
 
 if(NOT DEFINED SOURCE_DIR)
     message(FATAL_ERROR "SOURCE_DIR not defined")
 endif()
 
-set(CLIENT_DIR "${SOURCE_DIR}/src/client")
-set(OUTPUT_FILE "${CLIENT_DIR}/sources.cmake")
+if(NOT DEFINED TARGET_NAME)
+    message(FATAL_ERROR "TARGET_NAME not defined")
+endif()
+
+if(NOT DEFINED SEARCH_DIR)
+    message(FATAL_ERROR "SEARCH_DIR not defined")
+endif()
+
+set(TARGET_DIR "${SOURCE_DIR}/src/${TARGET_NAME}")
+set(OUTPUT_FILE "${TARGET_DIR}/sources.cmake")
 set(TEMP_FILE "${OUTPUT_FILE}.tmp")
 
 # Find all .cpp files relative to SOURCE_DIR
-file(GLOB_RECURSE CLIENT_CPP_FILES
+file(GLOB_RECURSE CPP_FILES
     RELATIVE "${SOURCE_DIR}"
-    "${CLIENT_DIR}/*.cpp"
+    "${SOURCE_DIR}/${SEARCH_DIR}/*.cpp"
 )
 
 # Create the temporary version of the output
 file(WRITE "${TEMP_FILE}" "# Auto-generated by generate_sources.cmake\n")
-file(APPEND "${TEMP_FILE}" "set(CLIENT_SOURCES\n")
+file(APPEND "${TEMP_FILE}" "set(${TARGET_NAME}_SOURCES\n")
 
-foreach(source_file IN LISTS CLIENT_CPP_FILES)
+foreach(source_file IN LISTS CPP_FILES)
     file(APPEND "${TEMP_FILE}" "  \${CMAKE_CURRENT_SOURCE_DIR}/${source_file}\n")
 endforeach()
