--- include/okami/data/maptype.cpp
@@ -1,200 +1,0 @@
-#include "maptype.hpp"
-
-#include <unordered_map>
-
-namespace okami
-{
-namespace MapTypes
-{
-static const char *names[MapTypes::NUM_MAP_TYPES] = {
-    "None",
-    "Kamiki Village (Game Start)",
-    "Cave of Nagi",
-    "Kamiki Village",
-    "Hana Valley",
-    "Tsuta Ruins",
-    "City Checkpoint",
-    "Tsuta Ruins (Spider Queen Arena)",
-    "Gale Shrine",
-    "Kusa Village",
-    "Sasa Sanctuary",
-    "Agata Forest (Madame Fawns House)",
-    "Digging Minigame",
-    "Onigiri Dojo (Lesson Room)",
-    "Gale Shrine (Crimson Helm Arena)",
-    "Calcified Cavern",
-    "Moon Cave (Interior)",
-    "Moon Cave (Staircase and Orochi Arena)",
-    "Ryoshima Coast (Bandit Spider Arena)",
-    "Taka Pass (Stray Bead 26 Cave)",
-    "Kamui (Cherry Bomb 3 Cave)",
-    "N Ryoshima Coast (Power Slash 2 Cave)",
-    "N Ryoshima Coast (Cherry Bomb 2 Cave)",
-    "Kamui (Ezofuji Power Slash 3 Cave)",
-    "Kamui (Blockhead Grande Cave)",
-    "N Ryoshima Coast (Bandit Spider Arena)",
-    "Kamui (Bandit Spider Arena)",
-    "Dragon Palace (Stray Bead 70 Cave)",
-    "N Ryoshima Coast (Stray Bead 63 Cave)",
-    "New Game Intro Cutscene Map",
-    "River of the Heavens",
-    "Seian City (Aristocratic Qtr)",
-    "Seian City (Commoners Qtr)",
-    "Seian City (Aristocratic Qtr - Himikos Palace)",
-    "Dragon Palace",
-    "Inside the Dragon",
-    "Sunken Ship",
-    "Imperial Palace (Ammy Size)",
-    "Imperial Palace (Issun Size)",
-    "Oni Island (Lower Interior)",
-    "Oni Island (Ninetails Arena)",
-    "Catcall Tower",
-    "Imperial Palace (Blight Arena)",
-    "Seian City (Commoners Qtr - Kimono Shop)",
-    "Oni Island (Exterior)",
-    "Oni Island (Upper Interior)",
-    "Oni Island (Sidescroller)",
-    "Wepkeer",
-    "Kamiki Village (100 years ago)",
-    "Wawku Shrine",
-    "Wawku Shrine (Nechku Arena)",
-    "Ponctan",
-    "Moon Cave (100 years ago)",
-    "Ark of Yamato",
-    "Ark of Yamato (Spider Queen Arena)",
-    "Ark of Yamato (Orochi Arena)",
-    "Ark of Yamato (Blight Arena)",
-    "Ark of Yamato (Ninetails Arena)",
-    "Ark of Yamato (Crimson Helm Arena)",
-    "Ponctan (Mrs. Seals House)",
-    "Yoshpet",
-    "Inner Yoshpet",
-    "Ark of Yamato (Yami Arena)",
-    "Wepkeer Square",
-    "Wawku Shrine (Lechku and Nechku Arena)",
-    "Fishing with Benkei (on Seian Bridge)",
-    "Fishing with Benkei (next to Himikos Palace)",
-    "Fishing with Kokariin (Agata)",
-    "Fishing with Kokariin (Kamui)",
-    "Unknown70",
-    "Fishing with Fisherman (in N Ryo)",
-    "Shinshu Field",
-    "Agata Forest",
-    "Moon Cave Entrance",
-    "Taka Pass",
-    "Ryoshima Coast",
-    "Unknown76",
-    "N Ryoshima Coast",
-    "Kamui",
-    "Kamui Ezofuji",
-    "Shinshu Field (100 years ago)",
-    "Moon Cave Entrance (100 years ago)",
-    "Unk82",
-};
-
-const char *GetName(unsigned value)
-{
-    if (value < MapTypes::NUM_MAP_TYPES)
-    {
-        return names[value];
-    }
-    return "invalid";
-}
-
-const std::unordered_map<unsigned, unsigned> mapidToIndex{
-    {0x100, KamikiVillageGameStart},
-    {0x101, CaveOfNagi},
-    {0x102, KamikiVillage},
-    {0x112, KamikiVillage},
-    {0x103, HanaValley},
-    {0x104, TsutaRuins},
-    {0x105, CityCheckpoint},
-    {0x106, TsutaRuinsSpiderQueenArena},
-    {0x107, GaleShrine},
-    {0x108, KusaVillage},
-    {0x109, SasaSanctuary},
-    {0x10A, AgataForestMadameFawnsHouse},
-    {0x10B, DiggingMinigame},
-    {0x10C, OnigiriDojoLessonRoom},
-    {0x10D, GaleShrineCrimsonHelmArena},
-    {0x10E, CalcifiedCavern},
-    {0x110, MoonCaveInterior},
-    {0x111, MoonCaveStaircaseandOrochiArena},
-    {0x113, RyoshimaCoastBanditSpiderArena},
-    {0x114, TakaPassStrayBead26Cave},
-    {0x115, KamuiCherryBomb3Cave},
-    {0x116, NRyoshimaCoastPowerSlash2Cave},
-    {0x117, NRyoshimaCoastCherryBomb2Cave},
-    {0x118, KamuiEzofujiPowerSlash3Cave},
-    {0x119, KamuiBlockheadGrandeCave},
-    {0x11A, NRyoshimaCoastBanditSpiderArena},
-    {0x11B, KamuiBanditSpiderArena},
-    {0x11C, DragonPalaceStrayBead70Cave},
-    {0x11D, NRyoshimaCoastStrayBead63Cave},
-    {0x120, NewGameIntroCutsceneMap},
-    {0x122, RiverOftheHeavens},
-    {0x200, SeianCityAristocraticQtr},
-    {0x201, SeianCityCommonersQtr},
-    {0x202, SeianCityAristocraticQtrHimikosPalace},
-    {0x203, DragonPalace},
-    {0x204, InsidetheDragon},
-    {0x205, SunkenShip},
-    {0x206, ImperialPalaceAmmySize},
-    {0x207, ImperialPalaceIssunSize},
-    {0x208, OniIslandLowerInterior},
-    {0x209, OniIslandNinetailsArena},
-    {0x20A, CatcallTower},
-    {0x20B, ImperialPalaceBlightArena},
-    {0x20C, SeianCityCommonersQtrKimonoShop},
-    {0x20D, OniIslandExterior},
-    {0x20E, OniIslandUpperInterior},
-    {0x20F, OniIslandSidescroller},
-    {0x301, Wepkeer},
-    {0x302, KamikiVillage100yearsago},
-    {0x303, WawkuShrine},
-    {0x304, WawkuShrineNechkuArena},
-    {0x305, Ponctan},
-    {0x306, MoonCave100yearsago},
-    {0x307, ArkOfYamato},
-    {0x308, ArkOfYamatoSpiderQueenArena},
-    {0x309, ArkOfYamatoOrochiArena},
-    {0x30A, ArkOfYamatoBlightArena},
-    {0x30B, ArkOfYamatoNinetailsArena},
-    {0x30C, ArkOfYamatoCrimsonHelmArena},
-    {0x30D, PonctanMrsSealsHouse},
-    {0x310, Yoshpet},
-    {0x311, InnerYoshpet},
-    {0x312, ArkOfYamatoYamiArena},
-    {0x313, WepkeerSquare},
-    {0x314, WawkuShrineLechkuAndNechkuArena},
-    {0xE00, FishingwithBenkeionSeianBridge},
-    {0xE01, FishingwithBenkeinexttoHimikosPalace},
-    {0xE02, FishingwithKokariinAgata},
-    {0xE03, FishingwithKokariinKamui},
-    {0xE04, FishingwithFishermaninNRyo},
-    {0xF01, ShinshuField},
-    {0xF02, ShinshuField},
-    {0xF03, AgataForest},
-    {0xF04, AgataForest},
-    {0xF06, MoonCaveEntrance},
-    {0xF07, TakaPass},
-    {0xF08, TakaPass},
-    {0xF09, RyoshimaCoast},
-    {0xF0A, RyoshimaCoast},
-    {0xF0C, NRyoshimaCoast},
-    {0xF11, Kamui},
-    {0xF12, Kamui},
-    {0xF13, KamuiEzofuji},
-    {0xF20, ShinshuField100yearsago},
-    {0xF21, MoonCaveEntrance100yearsago},
-};
-unsigned FromMapId(unsigned id)
-{
-    if (mapidToIndex.contains(id))
-    {
-        return mapidToIndex.at(id);
-    }
-    return None;
-}
-} // namespace MapTypes
-} // namespace okami
