+++ src/client/sources.cmake
@@ -1,5 +1,5 @@
 # Auto-generated by generate_sources.cmake
-set(CLIENT_SOURCES
+set(client_SOURCES
   ${CMAKE_CURRENT_SOURCE_DIR}/src/client/archipelagosocket.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/src/client/checks.cpp
   ${CMAKE_CURRENT_SOURCE_DIR}/src/client/console.cpp
