+++ src/library/sources.cmake
@@ -1,0 +1,14 @@
+# Auto-generated by generate_sources.cmake
+set(library_SOURCES
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/library/animals.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/library/bestiarytome.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/library/dataverify.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/library/dojotech.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/library/fish.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/library/maptype.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/library/memorymap.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/library/movelisttome.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/library/straybeads.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/library/travelguides.cpp
+  ${CMAKE_CURRENT_SOURCE_DIR}/src/library/treasures.cpp
+)
